refactor: build per-service proto durations in the service module
`nominal.ts` is transport-agnostic, but it was importing `nominal.protos.event.v2`
so it could build an `event_pb2.Duration`. Each service declares its own
`Duration` message — `nominal.event.v2.Duration`, `nominal.asset.v2.Duration`,
and a shared `nominal.types.Duration` — all structurally identical and none
assignable to another, so a single generically-named encoder in `nominal.ts`
cannot serve more than one of them.

`_to_proto_duration` and `_from_proto_duration` move to `nominal/core/event.py`,
which owns that service. `nominal.ts` keeps `_to_seconds_nanos_duration`, which
is the genuinely shared part and returns a plain tuple, and `_to_api_duration`
for the conjure type. `nominal.ts` now imports only shared `types/*` protos.

No behavior change: the split, the conjure encoder and the proto encoder all
still produce the same pair.
